### 2018-09-24 #977 @whedon
[PRE REVIEW]: TensorFlow.jl: An Idiomatic Julia Front End for TensorFlow
**Submitting author:** @malmaud (<a href="http://orcid.org/0000-0002-5352-2086">Jonathan Matthew Malmaud</a>)
**Repository:** <a href="https://github.com/malmaud/TensorFlow.jl" target ="_blank">https://github.com/malmaud/TensorFlow.jl</a>
**Version:** v0.10.1
