### 2018-09-25 #977 @oxinabox
I suggest @jekbradbury  might be a good reviewer; though he's not reviewed for JOSS before.
He is familiar with julia, and is on the Google TensorFlow team.
But has never contributed to TensorFlow.jl.
