+++ src/protobufs/worker_service_pb.jl
@@ -1,0 +1,6 @@
+# syntax: proto3
+using Compat
+using ProtoBuf
+import ProtoBuf.meta
+import ._ProtoBuf_Top_.tensorflow
+
