### 2018-10-06 #977 @jekbradbury
Sure, I'd be happy to.
