### 2018-10-09 4babb3c
=colon should never have been declared as an op in the first place
