### 2018-10-19 #1002 @jekbradbury
Accept.

Looks good to me. A couple quibbles (that don't block acceptance):
- The paper has a minor formatting issue where the table at the bottom of the first page overlaps with margin elements.
- I think this sentence should be dropped or reworded: "For example, it allows Julia code to operate on TPUs by virtue of using the same TensorFlow graph syntax as Python's TensorFlow client, even though there is no native Julia TPU compiler." This is because I don't believe anyone has actually demonstrated TensorFlow.jl code running on a TPU (and I believe significant external work would be needed to do this, because the TPU API is more than just TensorFlow graphs; you also have to use infeed/outfeed and explicitly invoke XLA compilation etc.)...and also because the last clause in the sentence may not be true for much longer :slightly_smiling_face:.
