### 2018-10-20 9e7cdb2
Paper revisions.
