### 2018-10-20 #1002 @malmaud
Thanks @jekbradbury, I agree with all your points. I've removed the TPU clause and fixed the table formatting.
