### 2018-10-22 #1002 @arokem
@jekbradbury : looks like most/all the issues have been addressed? Is this paper ready to be accepted from your point of view?

@malmaud: comments from me:

- [ ] In the third paragraph of the summary section, you have a sentence: "This * can correspond to ...". I believe this should be "Thus, * can correspond...". Do I understand correctly?
- [ ] The first sentence in the fourth paragraph might be missing a word. Maybe "TensorFlow.jl aims to..."
- [ ] The word "indices" is mis-spelled as "indicies" in the bottom of page 2
- [ ] The word "paramterized" is mis-spelled as "paramatrised"
- [ ] In the first paragraph of the "Challenges" section, you have the fragment: "... and does not include many conveients for the definition of training of graphs". How about replacing that with: "... and does not include interfaces for conveniently defining the training of graphs".  Is that statement true, by the way? Seems unlikely.
- [ ] The second paragraph of this section includes a couple of typos: "Graidnet" and "ptimizers".
- I still see a statement about TPUs in the manuscript currently on `master` in the last paragraph. I think this statement should at least be qualified (or better, removed).
