### 2018-10-22 #1002 @malmaud
Thanks for these comments!

* I fixed the typos you pointed out.
* The intent of the 'challenges' section is that TensorFlow's C API exposes minimal low-level functionality, with most of the complexity of building and training a neural net implemented in Python on top of the C API. I clarified the wording.
* I removed all references to TPUs. While I do think it would be pretty easy to get TensorFlow.jl to run on TPUs, it isn't implemented yet so I agree that removal makes sense.
