### 2018-10-22 #1002 @jekbradbury
@arokem: Yes, the paper's ready to be accepted from my POV.
