### 2018-10-26 5836e54
Add equality overrides for `Device` and `DevicePart`
This is needed for our XLA work, so that we can ensure that the device a
tensor is coming fron matches the device a model is running on.
