### 2018-10-29 #1002 @arokem
@jekbradbury : thank you for the review! 👏

@malmaud : the paper is ready to be accepted! To go ahead with the archiving of the paper, please create a DOI for the software (e.g., by tagging a release and uploading it to zenodo.org). Once you have a DOI, we can go ahead with the rest of the process.
