### 2018-10-29 7c030f7
Use AutoHashEquals.jl
