### 2018-10-31 #1002 @malmaud
OK, I have a DOI: 10.5281/zenodo.1476107. Thanks @jekbradbury and @arokem for handling the review.
