### 2018-11-01 #1002 @oxinabox
Cool, nice.
I am only finding put about thus now, as I guess the JOSS script doesn't tag coauthors.
