### 2018-11-04 #1002 @oxinabox
@arokem  I  was sharing a link around to the released paper,
and some errors were pointed out to me.

 - the citation for the MLandPL paper broke and so it didn't end up in the reference list.
 - the cited article for the Julia language is the 2014 preprint, rather than the final 2017 article
 - and a couple of typos slipped through.

What should we do?
Does the JOSS process allow us to submit an updated version?
