### 2018-11-04 #1002 @arfon
@oxinabox - feel free to update your paper in `master` on https://github.com/malmaud/TensorFlow.jl and check that it's fixed by typing `@whedon generate pdf` here.

Once you're happy with the changes, I can update the final accepted version.
