### 2018-11-05 #1002 @arfon
Paper compilation is failing for some reason right now. I've compiled locally:
[10.21105.joss.01002.pdf](https://github.com/openjournals/joss-reviews/files/2549836/10.21105.joss.01002.pdf)
