### 2018-12-28 #1002 @arfon
@oxinabox does it look correct here? https://doi.org/10.21105/joss.01002
