### 2018-12-29 #1002 @oxinabox
It seems it does not quiet have the last fixes in
https://github.com/malmaud/TensorFlow.jl/pull/469/
