### 2018-12-31 #1002 @arfon
> It seems it does not quiet have the last fixes in malmaud/TensorFlow.jl#469

OK, I've just regenerated (and updated) the live paper.
